refactor bar&butt
