+++ NetWorld-Immersion/NetWorld-Immersion.vcxproj
@@ -163,8 +163,10 @@
     <Image Include="..\Debug\UpperArmor__inventory_butt.png" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Bar.h" />
     <ClInclude Include="BattleGame.h" />
     <ClInclude Include="bmp.h" />
+    <ClInclude Include="Button.h" />
     <ClInclude Include="enemy.h" />
     <ClInclude Include="Inventory.h" />
     <ClInclude Include="LootGame.h" />
