+++ NetWorld-Immersion/main.cpp
@@ -61,96 +61,8 @@ DWORD AttackTime = 1000;
 DWORD AttackcurrentTime = 0;
 
 //TODO кнопки расположены абсолютно хаотично, если сделать это менее вырвиглазно и более структурированно и писать игру и рефакторить станет легче, как будто бы на стадии написани€ дл€ удобства можно отрисовать текст под нашими картиночками
-class Button {
-public:
-
-    float x, y, width, height;
-    HBITMAP hBitmap;//хэндл к спрайту шарика
-    HBITMAP hBitmapGlow;
-
-    bool CheckCollisionMouse() {
-        return Mouse.x < x + width && Mouse.x > x && Mouse.y < y + height && Mouse.y > y;
-    }
-    void Load(const char* imagename, const char* imagenameglow, float x_, float y_, float w, float h) {
-        x = x_; y = y_;
-        hBitmap = LoadBMP(imagename);
-        hBitmapGlow = LoadBMP(imagenameglow);
-        height = h * window.height;
-        width = w * window.width;
-        x = window.width / 2 - width * x;
-        y = window.height / 2 + height * y;
-    }
-    void LoadInv(const char* imagename, float x_, float y_, float w, float h) {
-        x = x_; y = y_;
-        hBitmap = LoadBMP(imagename);
-        height = h * window.height;
-        width = w * window.width;
-        x = window.width / 2 - width * x;
-        y = window.height / 2 + height * y;
-    }
-
-    bool Show() {
-        bool pw_collision = CheckCollisionMouse();
-        float offset = (float)pw_collision * sinf(timeGetTime() * .008);
-
-        if ((currentTime < healStartTime + healTime) || (AttackcurrentTime < AttackStartTime + AttackTime))
-        {
-            offset = 0;
-        }
-
-        ShowBitmap(x, y + offset, width, height, pw_collision ? hBitmapGlow : hBitmap);
-        return pw_collision;
-    }
-    bool ShowInv() {
-        bool pw_collision = CheckCollisionMouse();
-        ShowBitmap(x, y, width, height, hBitmap);
-        return pw_collision;
-    }
-
-    bool CheckCollisionMouseHeal() {
-        if (Mouse.L_butt) {
-            if (Mouse.x < x + width && Mouse.x > x && Mouse.y < y + height && Mouse.y > y)
-            {
-                if (currentTime > healStartTime + healTime) //todo
-                {
-                    player.adjustHeal();
-                    return true;
-                }
-
-            }
-        }
-        return false;
-    }
-};
-
-class Bar { //done and todo
-public:
-    float x, y, width, height, Health, Shield;
-    HBITMAP hBitmapBack;
-    HBITMAP hBitmapFront;
-    void Load(const char* imagenameBack, const char* imagenameFront, float x_, float y_, float w, float h) {
-        x = x_; y = y_;
-        hBitmapBack = LoadBMP(imagenameBack);
-        hBitmapFront = LoadBMP(imagenameFront);
-        height = h * window.height;
-        width = w * window.width;
-        x = window.width / 2 - width * x;
-        y = window.height / 2 + height * y;
-    }
-    void ShowHealth(int Health) { //TODO как будто бы можно сделать менее вырвиглазно в самой игре, да и код€ру оптимизировать
-        ShowBitmap(x, y, width, height, hBitmapBack);
-        ShowBitmap(x, y, Health / (float)player.HealthMax * width, height, hBitmapFront);
-    }
-    void ShowShield(int Shield) {
-        ShowBitmap(x, y, width, height, hBitmapBack);
-        ShowBitmap(x, y, Shield / (float)player.ShieldMax * width, height, hBitmapFront);
-    }
-    bool CheckCollisionMouse() {
-        return Mouse.x < x + width && Mouse.x > x && Mouse.y < y + height && Mouse.y > y;
-    }
-};
-
-Bar Health_bar, HealthEnemy_bar, Shield_bar, ShieldEnemy_bar; //TODO ¬џ–¬»√Ћјјјј«
+#include "Button.h"
+#include "Bar.h"
 
 Button PrimWeapon, SpecWeapon, DestructiveWeapon, EnemyB, Exit, Heal_butt, Inventory_butt, Boots__inventory_butt;
 HBITMAP hBack;// хэндл дл€ фонового изображени€
